Refactored types into useSubmitJob.ts
